refactor(server): split routes/groups into per-concern submodules
